==> scripts/02_Create_Seurat_object.Rmd <==
---
title: "02_Create_Seurat_object"
output: html_document
---

load libraries
```{r}
require(Seurat)
```

Load RData object 
```{r}
rm(list=ls())
dir <- "/home/ubuntu/scell_lung_adenocarcinoma/"
# dir <- "/myVolume/scell_lung_adenocarcinoma/"
load(file = paste(dir,"Data_input/objects/S01_Data_and_metadata.RData", sep=""))
```

Find ERCC's, compute the percent ERCC, and drop them from the raw data.
```{r}
erccs <- grep(pattern = "^ERCC-", x = rownames(x = raw.data), value = TRUE)
percent.ercc <- Matrix::colSums(raw.data[erccs, ])/Matrix::colSums(raw.data)
ercc.index <- grep(pattern = "^ERCC-", x = rownames(x = raw.data), value = FALSE)
raw.data <- raw.data[-ercc.index,]
dim(raw.data)
```

Create the Seurat object with all the data (unfiltered)
```{r}
main_tiss <- CreateSeuratObject(counts = raw.data)
# add rownames to metadta 
row.names(metadata) <- metadata$cell_id
# add metadata to Seurat object 
main_tiss <- AddMetaData(object = main_tiss, metadata = metadata)
main_tiss <- AddMetaData(object = main_tiss, percent.ercc, col.name = "percent.ercc")
# Head to check
head(main_tiss@meta.data)
```

Save prefiltered Seurat object 
```{r}
save(main_tiss, file=paste(dir,"Data_input/objects/S02_Main_Seurat_object_prefilter.RData", sep=""))
# load(paste(dir,"Data_input/Main_Seurat_object_prefilter.RData", sep=""))
```

Calculate percent ribosomal genes and add to metadata
```{r}
ribo.genes <- grep(pattern = "^RP[SL][[:digit:]]", x = rownames(x = main_tiss@assays$RNA@data), value = TRUE)
percent.ribo <- Matrix::colSums(main_tiss@assays$RNA@counts[ribo.genes, ])/Matrix::colSums(main_tiss@assays$RNA@data)
main_tiss <- AddMetaData(object = main_tiss, metadata = percent.ribo, col.name = "percent.ribo")
main_tiss
```

Filter cells so that remaining cells have nGenes >= 500 and nReads >= 50000
```{r}
main_tiss_filtered <- subset(x=main_tiss, subset = nCount_RNA > 50000 & nFeature_RNA > 500)
main_tiss_filtered
```

Save filtered Seurat object
```{r}
save(main_tiss_filtered, file=paste(dir,"Data_input/objects/S02_Main_Seurat_object_filtered.RData", sep=""))
```
